Move verify_all_proofs to a testing module
